Refactor ray.serve code base
