+++ python/ray/experimental/serve/tests/test_actors.py
@@ -7,8 +7,7 @@ import pytest
 
 import ray
 from ray.experimental.serve import SingleQuery
-from ray.experimental.serve.examples.adder import ScalerAdder, VectorizedAdder
-from ray.experimental.serve.examples.counter import Counter, CustomCounter
+from ray.experimental.serve.example_actors import ScalerAdder, VectorizedAdder, Counter, CustomCounter
 from ray.experimental.serve.object_id import get_new_oid
 
 INCREMENT = 3
